Spinner

+++ lib/workflow/steps/mr/CreateMr.ts
@@ -7,7 +7,8 @@ import { WorkflowOptions, WorkflowStep } from "..";
 type Reads = { target_branch: string; source_branch: string; repo: Repo };
 type Writes = { skipped?: Dir };
 type Options = {
-  temporary_branch: (reads: Reads) => string;
+  temporary_branch?: (reads: Reads) => string;
+  title: (reads: Reads) => string;
 };
 
 export class CreateMr extends WorkflowStep<Reads, Writes, Options> {
@@ -19,10 +20,11 @@ export class CreateMr extends WorkflowStep<Reads, Writes, Options> {
   async run(ctx: ExecutionContext, { target_branch, source_branch, repo }: Reads) {
     const ignores = ctx.config.repos.merge?.ignores;
     if (ignores?.includes(repo.dir.name())) return {};
-    const temporary_branch = this.options?.temporary_branch({ target_branch, source_branch, repo });
-    if ((await repo.getBranches()).includes(temporary_branch)) return { skipped: repo.dir };
-    await repo.stash(async () => {
 
+    const temporary_branch = this.options?.temporary_branch?.({ target_branch, source_branch, repo });
+    if (temporary_branch && (await repo.getBranches()).includes(temporary_branch)) return { skipped: repo.dir };
+
+    await repo.stash(async () => {
       const { original_branch } = await repo.switchBranchIfExists(target_branch);
       if (original_branch === temporary_branch) await repo.deleteBranch(temporary_branch);
       await repo.pull(target_branch);
@@ -30,11 +32,11 @@ export class CreateMr extends WorkflowStep<Reads, Writes, Options> {
       await repo.switchBranchIfExists(source_branch);
       await repo.pull(source_branch);
 
-      await repo.createNewBranch(temporary_branch);
-      await repo.createMr(target_branch, { title: `Nivelacion ${source_branch} - ${target_branch}` });
+      if (temporary_branch) await repo.createNewBranch(temporary_branch);
+      await repo.createMr(target_branch, { title: this.options.title({ source_branch, target_branch, repo }) });
       const { switched } = await repo.switchBranchIfExists(original_branch);
       if (!switched) await repo.switchBranchIfExists("master");
-      await repo.deleteBranch(temporary_branch);
+      if (temporary_branch) await repo.deleteBranch(temporary_branch);
       return false;
     });
